[not verified] Convert JS Components Button to TypeScript
